Separate TableBody
